send feed back velocity as an int value

+++ include/littlebot_firmware/serialization.h
@@ -34,7 +34,7 @@
 typedef struct SerializationClass Serialization;
 
 typedef void (*ptfDecode)(Serialization *self, char *msg, float *left_vel, float *right_vel);
-typedef void (*ptfEncode)(Serialization *self, char *msg, float *left_vel, float *right_vel);
+typedef void (*ptfEncode)(Serialization *self, char *msg, uint32_t *left_vel, uint32_t *right_vel);
 
 
 //structure of function pointer
@@ -48,7 +48,7 @@ void SerializationConstruct (Serialization *self);
 
 void fcDecode (Serialization *self, char *msg, float *left_vel, float *right_vel);
 
-void fcEncode (Serialization *self, char *msg, float *left_vel, float *right_vel);
+void fcEncode (Serialization *self, char *msg, uint32_t *left_vel, uint32_t *right_vel);
 
 void FloaToStr(float num, char *str, int afterpoint);
